added missing integer conversion for trimmed recovery values

+++ Pages/BuildingMeasurementPage.xaml.cs
@@ -54,8 +54,8 @@ namespace IndoorCO2MapAppV2.Pages
                     && double.TryParse(rec.AdditionalDataByParameter.GetValueOrDefault("trimHigh"),
                         System.Globalization.NumberStyles.Float, System.Globalization.CultureInfo.InvariantCulture, out double tHigh))
                 {
-                    TrimSlider.LowerValue = Math.Clamp(tLow, TrimSlider.Minimum, TrimSlider.Maximum);
-                    TrimSlider.UpperValue = Math.Clamp(tHigh, TrimSlider.Minimum, TrimSlider.Maximum);
+                    TrimSlider.LowerValue = Math.Clamp((int)tLow, TrimSlider.Minimum, TrimSlider.Maximum);
+                    TrimSlider.UpperValue = Math.Clamp((int)tHigh, TrimSlider.Minimum, TrimSlider.Maximum);
                 }
             });
